Add files via upload
In the modern world of science and technology, Transportation system is an integral part of living. Having this with us gives us the feel of the most civilized creatures on the planet. Automobiles play a vital role in our daily life but like every other thing, with some positives there are negatives too. Road accidents are the major threat to human lives. Speed is the key factor responsible for many of the accidents. Therefore, there is a need to control all the accidents.One of the methods for accident detection was, manual accident detection, which relied on the passengers passing by the accident spot who notify the concerned emergency authorities for any safety measures to be taken but this method was having a loophole in it because its efficiency was not reliable because someone have to witness the accident also this method incorporates delays and inaccuracies due to witness’s expression problem. Despite of many safety measures and programs launched by the governmental as well as nongovernmental organizations to make people aware of safe driving concept, life claimed by these accidents are increasing at an alarming rate.